fix tui ellipsis color
